pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "Agent template using orca-agent-sdk"
requires-python = ">=3.12"
dependencies = [
    "orca-agent-sdk",
    "agno",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]